requirements.txt
black==21.11b1
click==8.0.3
dataclasses==0.8
Django==2.2
djangorestframework==3.12.4
flake8==4.0.1
gunicorn==20.1.0
importlib-metadata==4.8.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
pkg-resources==0.0.0
platformdirs==2.4.0
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
regex==2021.11.10
sqlparse==0.4.2
tomli==1.2.2
typed-ast==1.5.0
typing-extensions==4.0.0
zipp==3.6.0
